feat(perf): B2 wire-in + u_main idempotent + server-ensure short-circuit
5건 라운드 — hexa-lang 0637c51 (parse leak fix) 이후 void 재빌드
가능해져 VC-64~VC-68 골화:

- B2 wire-in (VC-64): hx_is_wide_cjk extern fn 연결.
  void_main.hexa 인라인 wide_pages/is_wide_cjk_mixed/is_wide_cjk 삭제
  → widths.c 단일 source of truth. long ABI로 hexa int 호출 정합.
  너비·이모지 테이블 편집 45분→2.8초. bench 307M/s random /
  833M/s clustered @-O2, 18/18 self-check.

- u_main 이중호출 fix (VC-65): hexa-lang 0637c51 이후 hexa가
  u_main()을 자체 emit. build_void.sh의 기존 `grep -c < 2` 가드가
  못 감지하고 또 주입 → u_main 2회 실행으로 두 번째 run의
  scr_init이 첫 run에서 쓴 셀을 덮어써 타이핑 결과 소실. 멱등
  파이썬 패처 3-way (A: 이미 패치 / B: hexa standalone 승격 /
  C: legacy 주입).

- T28 COLS reset 누락 fix (VC-66): T28이 alt-screen 테스트 위해
  COLS=10/ROWS=3 설정 후 reset 없음. T45 (TAB 15→16) T46 (TAB
  77→79) T50 (VPA row 4) CUP/VPA 좌표가 COLS-1=9로 clamp 당해
  실패. T45 진입 시 80/24/1920 복원. 53/28 all green.

- server_ensure 1초 지연 제거 (VC-67): sample 2102건 중 970건
  (46%)이 applicationWillTerminate/startup의 void_server_list →
  void_server_ensure → spawn_daemon → usleep×100(최대 1초)에
  박힘. try_connect 먼저, 소켓 없으면 skip. on-disk checkpoint
  정리는 서버 미기동 시에도 항상 수행.

- prompt_forward_keys 단순화 (VC-68): FEATURE-P0-1 per-byte 루프가
  printable을 line_buf에만 쌓고 PTY 미전송 → 화면 공백. 추가
  forward 주입 시 Enter 때 line_buf 재플러시로 'cl'→'clcl' 이중
  에코. 원래 thin wrapper `return hexa_keys_to_pty(master_fd)`
  복원. builtin(cd/pwd/clear/history/help) + ↑/↓ 히스토리
  인터셉트 임시 비활성 — shell readline 위임.

BUILD_SPEED_PLAN: B2 ✅ (draft→landed), B6 드롭 (clang 단계 1%라
ROI 최저, hexa-lang separate compilation 등 업스트림 우선).

pending (convergence.json pending_ossification):
  partial-redraw 깜빡임 — setNeedsDisplay:YES 전체-뷰 폴백이
  macOS 15+ backing-store 미보존 workaround. setNeedsDisplayInRect
  + layerContentsRedrawPolicy 교체 1차 시도는 TERM_PAD y-offset
  누락으로 롤백. 정확한 offset + 회귀 테스트 후 재도전.

+++ src/widths.c
@@ -68,7 +68,7 @@ enum { WIDE_BMP_MIXED_N = (int)(sizeof(wide_bmp_mixed) / sizeof(wide_bmp_mixed[0
 // Small n (10) — linear scan beats binary search (branch-predictable,
 // all ranges fit in one cache line). Caller already knows cp is in a
 // "mixed" page, so we skip the initial `cp < first_range` shortcut.
-static int is_wide_cjk_mixed(int cp) {
+static long is_wide_cjk_mixed(long cp) {
     for (int i = 0; i < WIDE_BMP_MIXED_N; i++) {
         if (cp < wide_bmp_mixed[i][0]) return 0;
         if (cp <= wide_bmp_mixed[i][1]) return 1;
@@ -81,16 +81,16 @@ static int is_wide_cjk_mixed(int cp) {
 // then BMP CJK, then SMP/emoji. Reordering away from hexa source is
 // intentional and semantically identical (see anchor tests in
 // tests/bench_width.c).
-int hx_is_wide_cjk(int cp) {
+long hx_is_wide_cjk(long cp) {
     // Fast path: ASCII / Latin / common scripts — vast majority of
     // real terminal input. Matches hexa `if cp < 4352 { return 0 }`.
-    if ((unsigned)cp < 4352u) return 0;
+    if ((unsigned long)cp < 4352ul) return 0;
 
     // BMP O(1) page-table lookup. This is the hot branch for mixed
     // CJK/Latin content — ~30% of realistic terminal streams.
-    if ((unsigned)cp <= 65535u) {
-        uint8_t tag = wide_pages[(unsigned)cp >> 8];
-        if (tag < 2) return (int)tag;   // 0 → narrow, 1 → wide
+    if ((unsigned long)cp <= 65535ul) {
+        uint8_t tag = wide_pages[(unsigned long)cp >> 8];
+        if (tag < 2) return (long)tag;   // 0 → narrow, 1 → wide
         return is_wide_cjk_mixed(cp);
     }
 
@@ -108,27 +108,19 @@ int hx_is_wide_cjk(int cp) {
 
 // ── Emoji helpers (forward-compatible; not yet called from hexa) ────
 
-// Fitzpatrick skin-tone modifiers (U+1F3FB EMOJI MODIFIER FITZPATRICK
-// TYPE-1-2 .. U+1F3FF EMOJI MODIFIER FITZPATRICK TYPE-6).
-int hx_is_emoji_modifier(int cp) {
+long hx_is_emoji_modifier(long cp) {
     return (cp >= 0x1F3FB && cp <= 0x1F3FF) ? 1 : 0;
 }
 
-// ZERO WIDTH JOINER — glues emoji sequences into a single grapheme.
-int hx_is_zwj(int cp) {
+long hx_is_zwj(long cp) {
     return (cp == 0x200D) ? 1 : 0;
 }
 
-// VARIATION SELECTOR-16 — forces emoji presentation on dual-use cp.
-int hx_is_vs16(int cp) {
+long hx_is_vs16(long cp) {
     return (cp == 0xFE0F) ? 1 : 0;
 }
 
-// Supplementary Multilingual Plane CJK (Ext-B..G, as classified in
-// void_main.hexa:594–595). Kept separate from hx_is_wide_cjk so
-// callers can distinguish "wide because SMP CJK" from "wide because
-// emoji". Matches the hexa source's two back-to-back range checks.
-int hx_is_smp_cjk(int cp) {
+long hx_is_smp_cjk(long cp) {
     if (cp >= 131072 && cp <= 196605) return 1;   // Ext-B..F
     if (cp >= 196608 && cp <= 262141) return 1;   // Ext-G
     return 0;
